// src/app/models/storage-values.ts
/**********************************************************************
Name: storage-values.ts
Copyright © 2024 

Purpose:
This Model will be the enum used to make the calls of the storage.
Reason why i did this was to make a code a bit more clean in terms of strings.

History
VERSION     AUTHOR    		  DATE 		      DETAIL					
1.0    	Ricardo Imperial 29/03/2024         Model Class.		
***********************************************************************/
export enum StorageEnum{
  DEFAULT = 'default',
  COLOR = 'color',
  COLORDESCRIPTION = 'colordesc',
  COLORCOST = 'colorprice',
  MODEL= 'model',
  MODELDESCRIPTION= 'modeldesc',
  CONFIG = 'config',
  CONFIGDESCRIPTION = 'configdesc',
  CONFIGCOST = 'configprice',
  RANGE = 'range',
  SPEED = 'speed',
  HITCH = 'hitch',
  STEERING = 'steering',
};